refactor: review fixes
